rename folder and file
