+++ packages/plugins/auto-doc/src/components/doc-template-container/doc-template-container.svelte
