+++ packages/plugins/auto-doc/src/components/index.ts
@@ -1,1 +1,1 @@
-export { default as ElementsTypeContainer } from './elements-type-container/element-type-container.svelte';
+export { default as ElementsTypeContainer } from './doc-template-container/doc-template-container.svelte';
